requirements.txt
fastapi
uvicorn
jinja2
pandas
pyodbc
python-dotenv
python-multipart
